Fix a type error in the TranslationsList component
There was a type error in the filterSourceMessages() method of the TranslationsList.vue component that several users complained about. As it turned out sourceMessage.languages[language.id] can sometimes be undefined but in the if statement it was only compared to null which caused the if statement to evaluate to true because undefined is not null. Which in its turn caused a call to the toLowerCase() method of undefined.